Numerous changes relating to supporting new sprites.
game.py:

* Renumbered all sprites except for EMPTY to be high non-alpha
characters. They are now (more or less) the last row of the font page.

* Since there are three different sprites for the invaders, I also
created an array, INVADER_SPRITE that has the sprite characters
indexed for easier redrawing by invader type (see note on Invader.py,
below).

* The PLAYER now has three different sprites: left, right and center of
the tank. Changes throughout to support checking for the different
sprites (instead of just '@').

* Changed how invaders are laid out. 5 rows of 11 now, spaced every
other column and every other row. The invader generated depends on the
row, and when the Invader() is instantiated, the sprite number is set. I
originally called this "type" but it is a reserved word. It might make
sense to change score based on what type of invader is hit.

* Collision detection now has to check for different types of invaders.

* Redrawing has to erase multiple types of invaders. There has to be a
better way.

* Invaders are redrawn based on their "sprite" property using the
INVADER_SPRITE list.

Invader.py:

* Added the sprite type.

Also committed is an updated font PNG and the XCF source (gimp)
including an alignment grid (make this invisible when exporting).