Fixes #5 -- problem of games crashing randomly
The reason this was happening is that the game did not handle the case where the player actually killed all the invaders! It tried to move the invaders down a row, but there were no invaders to move.

I fixed this by calling draw_level to reset the level. Since we redrew the level, I had to set mothership_exists to false.

This uncovered another problem, however, which is that once the map is redrawn, the mothership could appear at any time (due to turn count) but the invaders start in the top row, so they can conflict, causing an issue where mothership_exists is true, but certain mothership tiles aren't present because the invaders overwrote them.

I fixed this by moving the invaders' starting position down a row.

Working on this uncovered another bug, however, which is that the invaders physically colliding with the ship do not kill it, even though the code seems to be written to catch this case. Players don't die until the invader hits the edge of the screen in the bottom row. A new bug will be issued shortly.